% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backward_selection.R
\name{BWselection}
\alias{BWselection}
\title{BackWard Selection}
\usage{
BWselection(
  data,
  qvarlist = NULL,
  lvarlist = NULL,
  spvarlist = NULL,
  spclist = NULL,
  catvarlist = NULL,
  outcome,
  type = "lm",
  sig = 0.05,
  complete_case = TRUE,
  track = "Simple"
)
}
\arguments{
\item{data}{input Data used for the regression}

\item{qvarlist}{Variable list in vector format, including all the variables needed for the quadratic form}

\item{lvarlist}{Variable list in vector format, including all the variables in the continuous linear form}

\item{spvarlist}{variable list in vector format, including all the variables for the spline form}

\item{spclist}{if spvarlist is not empty, specify the outpoints for the spline term}

\item{catvarlist}{variable list in the vector format, including all the categorical variables}

\item{outcome}{Outcome variable for the regression model}

\item{type}{type of regression model, default is lm, logistic regression use type="logistic"}

\item{sig}{significant level used for exit the model, default is 0.05}

\item{complete_case}{whether to use the complete case for all the variables in the model, default is TRUE}

\item{track}{Whether print the selection process, could select from "Simple","All", "None"}
}
\value{
a list of original formula, final fit model and final fit formula
}
\description{
Perform Backward selection on regression models
}
\examples{

#Load NHANES DATA
library(NHANES)
data("NHANES")
data=NHANES

#Created variable lists for the model selection
qvarlist<-c("BPSysAve","SleepHrsNight","TotChol")
lvarlist<-c("Age","BPDiaAve","Weight","Height")
spvarlist<-c("Pulse","DirectChol")
spclist<-c(70,1.2)
catvarlist<-c("Depressed","Marijuana","Gender")

#Run the model selection, this will return a list
fit_test1=BWselection(data=NHANES,qvarlist=qvarlist,lvarlist=lvarlist,spvarlist=spvarlist,
                     spclist=spclist,catvarlist=catvarlist,outcome="BMI",type="lm",sig=0.05,complete_case=TRUE)
}
